Update SQL Basics Simple JOIN and RANK.sql
I learned the RANK() function; never knew about it! Postgre is full of mysteries. I'm so used to functions requiring a parameter to be passed inside the function (in SQL, of course Python has plenty of functions that don't require a parameter) so that was new for me. I don't fully understand how you actually use two functions, the second being OVER(), in order to create a single column, so I'll have to investigate that further.