Add files via upload
Similar to "Simple IN" but adding an extra layer of complication by using a WITH clause to give a name to the subquery (which calls another table), then calling the specific column from the subquery in the IN operator. One thing that surprised me is that while the subquery in my answer has only one column (id), you still have to specify the column which the IN operator will use as a comparison against the column in the SELECT statement's WHERE clause.