Add files via upload
Initially this seemed like a simple problem, but I used the CASE SQL function, and the result was separate rows for Mike and Jon's aggregate functions. That means that in rows where the result for Jon's aggregate function was returned, Jon's columns would have a NULL value, and vice versa. I discovered the Postgresql FILTER function and that allowed me to return the aggregate function output into the same rows, which both allowed me to actually solve the problem, as well as have a more readable output. In the future I'll look into how to do that in other SQL variants.