Create New-Companies.sql
Problem can be found here: https://www.hackerrank.com/challenges/the-company/problem

The main lesson here was to remember to be careful of columns that don't *seem* to need a group by clause, by default MySQL will throw an error if non-functionally dependent columns are not grouped when other columns are using group by functions. In this case the "Founder" column needed to be grouped, with my initial solution I did not do that.